git: Ignore some Java files in git archive
When used for building a package, (Debian, rpm, etc), certain Java and
NetBeans files must be omitted, otherwise the package build will fail.
They should still be under version control however. The solution is to
specify that they should be ignored by the git archive command which is
used to produce source tarballs.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>