m4: Add Sparse-dependent CPPFLAGS
Sparse's knowledge of gcc attributes is often out of date and for long
periods. Switch off unknown attribute warnings.

gcc and sparse ignore unrecognised -W parameters so this could have been
implemented with complete disregard to a) Whether sparse was requested,
and b) The version of sparse needed for an action. Request for sparse
and sparse version checking are implemented none the less as it makes
for a cleaner more logical output from make.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>